+++ python/alpha93/packages/commons/src/alpha93/commons/types.py
@@ -11,5 +11,5 @@ def any_object(obj: object | None = None, /):
 
 class typed[T]:
     @staticmethod
-    def getattr(self, name: str, default = ...) -> T:  # ruff: ignore[bad-staticmethod-argument]
+    def getattr(self, name: str, default = ...) -> T:  # noqa: PLW0211
         return getattr(self, name) if default is ... else getattr(self, name, default)